[fixed] constant values in idl translation [fixed] some warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1645 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/scene/LightUpdateTransformMessage.h
@@ -67,7 +67,7 @@ public:
 		insertByte(0); // unknown
 		
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte) scno->getDirectionAngle());
 		
 		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 			 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << "," 
@@ -90,7 +90,7 @@ public:
 		insertByte(0); // unknown
 
 		// add direction
-		insertByte(scno->getDirectionAngle());
+		insertByte((byte) scno->getDirectionAngle());
 
 		/*System::out << "Position Update [" << player->getObjectID() << "] (" 
 				 << (int) (player->getPositionX()) << "," << (int) (player->getPositionZ()) << ","
